# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories:
    get:
      operationId: getStories
      tags: [stories]
      summary: Get stories from people you follow
      responses:
        "200":
          description: Stories grouped by user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryGroup"
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a new story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{storyId}/view:
    post:
      operationId: viewStory
      tags: [stories]
      summary: Mark a story as viewed
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewResult"

  # ─── MESSAGES ────────────────────────────────────────────────────────────────
...
  /streams:
    get:
      operationId: getLiveStreams
      tags: [streams]
      summary: Get active live streams
      responses:
        "200":
          description: Live streams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
    post:
      operationId: startStream
      tags: [streams]
      summary: Start a live stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamInput"
      responses:
        "201":
          description: Created stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
  /streams/{streamId}:
    get:
      operationId: getStream
      tags: [streams]
      summary: Get a stream
      parameters:
        - name: streamId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
  /streams/{streamId}/end:
    post:
      operationId: endStream
      tags: [streams]
      summary: End a live stream
      parameters:
        - name: streamId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # ─── AI ──────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, creator, company, recruiter, influencer, moderator, admin]
        isVerified:
          type: boolean
        isPremium:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        experience:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
        socialLinks:
          type: array
          items:
            $ref: '#/components/schemas/SocialLink'
        createdAt:
          type: string
...
    Story:
      type: object
      required: [id, authorId, author, mediaUrl, mediaType, createdAt, expiresAt]
      properties:
        id:
          type: string
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/User"
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        text:
          type: ["string", "null"]
        viewsCount:
          type: integer
        isViewed:
          type: boolean
        createdAt:
          type: string
        expiresAt:
          type: string
    StoryGroup:
      type: object
      required: [user, stories, hasUnviewed]
      properties:
        user:
          $ref: "#/components/schemas/User"
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
        hasUnviewed:
          type: boolean
    StoryInput:
      type: object
      required: [mediaUrl, mediaType]
      properties:
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        text:
          type: string
    ViewResult:
      type: object
      required: [viewed]
      properties:
        viewed:
          type: boolean

    # ─── MESSAGE SCHEMAS ──────────────────────────────────────────────────────
...
    ActionResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── JOB SCHEMAS ─────────────────────────────────────────────────────────
...
    Stream:
      type: object
      required: [id, title, hostId, host, isLive, viewersCount, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        hostId:
          type: string
        host:
          $ref: "#/components/schemas/User"
        isLive:
          type: boolean
        viewersCount:
          type: integer
        peakViewers:
          type: integer
        category:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    StreamInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        category:
          type: string

    # ─── AI SCHEMAS ───────────────────────────────────────────────────────────
...
    Skill:
      type: object
      required: [id, skill, level]
      properties:
        id:
          type: string
        skill:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
...
    Experience:
      type: object
      required: [id, company, title]
      properties:
        id:
          type: string
        company:
          type: string
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        current:
          type: boolean
...
    Education:
      type: object
      required: [id, institution, degree]
      properties:
        id:
          type: string
        institution:
          type: string
        degree:
          type: string
        field:
          type: string
        startYear:
          type: string
        endYear:
          type: string
        current:
          type: boolean
...
    Language:
      type: object
      required: [id, language, proficiency]
      properties:
        id:
          type: string
        language:
          type: string
        proficiency:
          type: string
          enum: [basic, conversational, fluent, native]
...
    SocialLink:
      type: object
      required: [id, platform, url]
      properties:
        id:
          type: string
        platform:
          type: string
        url:
          type: string
...